# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fake News Detection API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ArticleInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 10 */
  body: string;
}

export type FeatureValueCategory = typeof FeatureValueCategory[keyof typeof FeatureValueCategory];


export const FeatureValueCategory = {
  linguistic: 'linguistic',
  structural: 'structural',
} as const;

export interface FeatureValue {
  name: string;
  label: string;
  value: number;
  importance: number;
  category: FeatureValueCategory;
}

export type ExperimentResultConfiguration = typeof ExperimentResultConfiguration[keyof typeof ExperimentResultConfiguration];


export const ExperimentResultConfiguration = {
  'linguistic-only': 'linguistic-only',
  'structural-only': 'structural-only',
  combined: 'combined',
} as const;

export type ExperimentResultPrediction = typeof ExperimentResultPrediction[keyof typeof ExperimentResultPrediction];


export const ExperimentResultPrediction = {
  fake: 'fake',
  real: 'real',
} as const;

export interface ExperimentResult {
  configuration: ExperimentResultConfiguration;
  prediction: ExperimentResultPrediction;
  confidence: number;
  featureCount: number;
}

export type CorroboratingSourceType = typeof CorroboratingSourceType[keyof typeof CorroboratingSourceType];


export const CorroboratingSourceType = {
  credible: 'credible',
  factcheck: 'factcheck',
} as const;

export interface CorroboratingSource {
  title: string;
  url: string;
  snippet: string;
  publisher: string;
  type: CorroboratingSourceType;
}

export interface FactCheckResult {
  claimant: string;
  claimText: string;
  claimDate?: string | null;
  rating: string;
  publisher: string;
  publisherSite: string;
  url: string;
  reviewDate?: string | null;
}

export type ClassificationResultPrediction = typeof ClassificationResultPrediction[keyof typeof ClassificationResultPrediction];


export const ClassificationResultPrediction = {
  fake: 'fake',
  real: 'real',
} as const;

export interface ClassificationResult {
  id: number;
  title: string;
  prediction: ClassificationResultPrediction;
  confidence: number;
  features: FeatureValue[];
  experiments: ExperimentResult[];
  explanation: string;
  createdAt: string;
}

export type HistoryItemPrediction = typeof HistoryItemPrediction[keyof typeof HistoryItemPrediction];


export const HistoryItemPrediction = {
  fake: 'fake',
  real: 'real',
} as const;

export interface HistoryItem {
  id: number;
  title: string;
  prediction: HistoryItemPrediction;
  confidence: number;
  createdAt: string;
}

export interface HistoryList {
  items: HistoryItem[];
  total: number;
}

export type FeatureMetaCategory = typeof FeatureMetaCategory[keyof typeof FeatureMetaCategory];


export const FeatureMetaCategory = {
  linguistic: 'linguistic',
  structural: 'structural',
} as const;

export interface FeatureMeta {
  name: string;
  label: string;
  description: string;
  category: FeatureMetaCategory;
}

export type ExperimentPerformanceConfiguration = typeof ExperimentPerformanceConfiguration[keyof typeof ExperimentPerformanceConfiguration];


export const ExperimentPerformanceConfiguration = {
  'linguistic-only': 'linguistic-only',
  'structural-only': 'structural-only',
  combined: 'combined',
} as const;

export interface ExperimentPerformance {
  configuration: ExperimentPerformanceConfiguration;
  featureCount: number;
  accuracy: number;
  precision: number;
  recall: number;
  f1Score: number;
  rocAuc: number;
  confusionMatrix: number[][];
}

export interface ModelPerformance {
  experiments: ExperimentPerformance[];
  trainingSamples: number;
  folds: number;
  featureImportances: FeatureValue[];
}

export type GetHistoryParams = {
limit?: number;
offset?: number;
};

export type GetCorroboratingSourcesParams = {
q: string;
prediction: GetCorroboratingSourcesPrediction;
};

export type GetCorroboratingSourcesPrediction = typeof GetCorroboratingSourcesPrediction[keyof typeof GetCorroboratingSourcesPrediction];


export const GetCorroboratingSourcesPrediction = {
  fake: 'fake',
  real: 'real',
} as const;

export type GetCorroboratingSources200 = {
  results: CorroboratingSource[];
};

export type GetFactChecksParams = {
q: string;
};

export type GetFactChecks200 = {
  results: FactCheckResult[];
};