# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fake News Detection API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: classify
    description: Fake news classification
  - name: history
    description: Classification history
  - name: model
    description: Model metadata and performance
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /classify:
    post:
      operationId: classifyArticle
      tags: [classify]
      summary: Classify a news article
      description: Preprocesses text, extracts 20 linguistic and structural features, and classifies using Random Forest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "200":
          description: Classification result with feature importance analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassificationResult"
        "400":
          description: Invalid input

  /history:
    get:
      operationId: getHistory
      tags: [history]
      summary: Get classification history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Past classifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryList"

  /history/{id}:
    get:
      operationId: getHistoryItem
      tags: [history]
      summary: Get a specific classification result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassificationResult"
        "404":
          description: Not found

  /corroborate:
    get:
      operationId: getCorroboratingSources
      tags: [classify]
      summary: Find corroborating sources via Google Custom Search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: prediction
          in: query
          required: true
          schema:
            type: string
            enum: [fake, real]
      responses:
        "200":
          description: Corroborating sources from credible outlets or fact-checkers
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/CorroboratingSource"

  /fact-check:
    get:
      operationId: getFactChecks
      tags: [classify]
      summary: Query Google Fact Check Tools API
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching fact-checks from verified publishers
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/FactCheckResult"

  /model/features:
    get:
      operationId: getFeatureMeta
      tags: [model]
      summary: Get feature descriptions
      responses:
        "200":
          description: All 20 features with descriptions and categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureMeta"

  /model/performance:
    get:
      operationId: getModelPerformance
      tags: [model]
      summary: Get model evaluation metrics
      description: Returns accuracy, precision, recall, F1-score, ROC-AUC, and confusion matrix from 10-fold cross-validation for all three experimental configurations
      responses:
        "200":
          description: Cross-validation performance metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPerformance"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ArticleInput:
      type: object
      required: [title, body]
      properties:
        title:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 10

    FeatureValue:
      type: object
      required: [name, label, value, importance, category]
      properties:
        name:
          type: string
        label:
          type: string
        value:
          type: number
        importance:
          type: number
        category:
          type: string
          enum: [linguistic, structural]

    ExperimentResult:
      type: object
      required: [configuration, prediction, confidence, featureCount]
      properties:
        configuration:
          type: string
          enum: [linguistic-only, structural-only, combined]
        prediction:
          type: string
          enum: [fake, real]
        confidence:
          type: number
        featureCount:
          type: integer

    CorroboratingSource:
      type: object
      required: [title, url, snippet, publisher, type]
      properties:
        title:
          type: string
        url:
          type: string
        snippet:
          type: string
        publisher:
          type: string
        type:
          type: string
          enum: [credible, factcheck]

    FactCheckResult:
      type: object
      required: [claimant, claimText, rating, publisher, publisherSite, url]
      properties:
        claimant:
          type: string
        claimText:
          type: string
        claimDate:
          type: string
          nullable: true
        rating:
          type: string
        publisher:
          type: string
        publisherSite:
          type: string
        url:
          type: string
        reviewDate:
          type: string
          nullable: true

    ClassificationResult:
      type: object
      required: [id, title, prediction, confidence, features, experiments, explanation, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        prediction:
          type: string
          enum: [fake, real]
        confidence:
          type: number
        features:
          type: array
          items:
            $ref: "#/components/schemas/FeatureValue"
        experiments:
          type: array
          items:
            $ref: "#/components/schemas/ExperimentResult"
        explanation:
          type: string
        createdAt:
          type: string

    HistoryItem:
      type: object
      required: [id, title, prediction, confidence, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        prediction:
          type: string
          enum: [fake, real]
        confidence:
          type: number
        createdAt:
          type: string

    HistoryList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/HistoryItem"
        total:
          type: integer

    FeatureMeta:
      type: object
      required: [name, label, description, category]
      properties:
        name:
          type: string
        label:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [linguistic, structural]

    ExperimentPerformance:
      type: object
      required: [configuration, featureCount, accuracy, precision, recall, f1Score, rocAuc, confusionMatrix]
      properties:
        configuration:
          type: string
          enum: [linguistic-only, structural-only, combined]
        featureCount:
          type: integer
        accuracy:
          type: number
        precision:
          type: number
        recall:
          type: number
        f1Score:
          type: number
        rocAuc:
          type: number
        confusionMatrix:
          type: array
          items:
            type: array
            items:
              type: integer

    ModelPerformance:
      type: object
      required: [experiments, trainingSamples, folds, featureImportances]
      properties:
        experiments:
          type: array
          items:
            $ref: "#/components/schemas/ExperimentPerformance"
        trainingSamples:
          type: integer
        folds:
          type: integer
        featureImportances:
          type: array
          items:
            $ref: "#/components/schemas/FeatureValue"