# lib/api-zod/src/generated/types/factCheckResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fake News Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface FactCheckResult {
  claimant: string;
  claimText: string;
  claimDate?: string | null;
  rating: string;
  publisher: string;
  publisherSite: string;
  url: string;
  reviewDate?: string | null;
}