# lib/api-zod/src/generated/types/featureMeta.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fake News Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { FeatureMetaCategory } from './featureMetaCategory';

export interface FeatureMeta {
  name: string;
  label: string;
  description: string;
  category: FeatureMetaCategory;
}