# lib/api-zod/src/generated/types/historyList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fake News Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryItem } from './historyItem';

export interface HistoryList {
  items: HistoryItem[];
  total: number;
}